+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to `once` will be documented in this file
 
+## 2.1.0 - 2019-06-19
+
+- add enabling/disable the cache
+- add `flush` method
+- drop support for PHP 7.1 and below
+
 ## 2.0.2 - 2018-12-27
 
 - add ability to use `once` in static functions